added profile for generating the distro
Added a profile and moved the creation of a distribution with the
maven-assembly-plugin to the profile.  The profile isn't active by
default.  Previously, the distribution was getting generated as part of
the deploy process, which caused problems with uploads to Sonatype
OSSRH.  The distribution file shouldn't be part of the deploy to Maven
Central.