Synchronize commit method
This filter uses the current HEAD revision to determine if the commit is out of sync. It then proceeds with
the rest of the commit. The rest of the commit involves synchronized access to the change management pool and
eventually writing to the history file. Geting the HEAD revision also involves synchronized access to the pool.
Two clients could both issue commits, and the second client call getHead before the first does the actual write.
Both would be able to proceed to write, corrupting the file. In the case where both commits refence the same start
revision, one steps on the other when the history file is read as the revisions are used to create a hash map on read.
To the clients it would appear the client B just committed first when in fact A did and got stepped on.