WIP - proof of concept for snaphots
When a project is created a snapshot of the ontology is serialized and pushed to the server. This is
stored rather than an initial huge change commit. When the project is opened this snapshot is retrieved
if it does not already exist on the client. Together with keeping the HEAD revision in the change pool
entry this results in significant (order of magnitude) speedups. A large 150M ontology was taking as long as
3 or 4 minutes for simple operations, mainly due to scanning the entire history file and de-serializing that
initial large commit.