Modify login service to support different services
The authentication class to use can be specified in the config file. The class needs to implement the LoginService.
The LoginService has two methods, login, that takes user name and password, and setConfig, in order to make use of the
ServerConfiguration. The config is likely needed in all authentication classes, in the default to validate the user and
her hashed password, in LDAP to get details needed from the config, etc...