Modified the DefaultLoginService to better fit the HTTP server
The SessionManager is pulled out, as adding the token to it is handled by the server.
This service requires a salted password and so was throwing an exception if a plain password was passed in. If a
plain password is passed the method now goes ahead and does the hashing, allowing the server to ignore these details.